+++ src/components/Navbar/Navbar.jsx
@@ -2,9 +2,6 @@ export const Navbar = () => {
     return(
         <nav className="">
             <h1 className="">socio<span>Flasher</span></h1>
-            <div className="">
-                <img className="" src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRkg3YrRehU5lvlYruB90y6K6cF4eyTsMk2iZhcG0N28iHyXrG0_VvlFeBxjcq3rhyl-Po&usqp=CAU" alt="user" />
-            </div>
         </nav>
     )
 }
